Break out the reducers and their tests into separate files.
