+++ assets/app/reducers/builds.js
@@ -1,0 +1,11 @@
+import { buildActionTypes } from "../constants";
+
+export function builds(state = [], action) {
+  switch (action.type) {
+  case buildActionTypes.BUILDS_RECEIVED:
+    return action.builds;
+  default:
+    return state;
+  }
+}
+
